+++ src/layout/Content.svelte
@@ -6,6 +6,7 @@
     import Notepad from '../apps/Notepad.svelte';
     import Whiteboard from '../apps/Whiteboard.svelte';
     import JWT from '../apps/jwt.svelte';
+    import Base64 from '../apps/Base64.svelte';
 
     const appMap: { appCode: string, entryComponent: SvelteComponent }[] = [
         { appCode: 'json_viewer', entryComponent: JsonViewer },
@@ -13,6 +14,7 @@
         { appCode: 'notepad', entryComponent: Notepad },
         { appCode: 'whiteboard', entryComponent: Whiteboard },
         { appCode: 'jwt', entryComponent: JWT },
+        { appCode: 'base64', entryComponent: Base64 },
     ];
 
     /**